Move LibvirtDomainManager to separate file
